backend-app-api: move http dir out of lib
